+++ caQtDM_Lib/caQtDM_Plugins/bsread/bsread_wfconverter.h
@@ -75,7 +75,7 @@ public:
         timer.start();
 
 
-        if (kDataP->edata.valueCount!=bsreadPVP->bsdata.wf_data_size){
+        if ((ulong)kDataP->edata.valueCount!=bsreadPVP->bsdata.wf_data_size){
             QMutex *datamutex;
             datamutex = (QMutex*) kDataP->mutex;
             datamutex->lock();
